+++ zstd_cffi.py
@@ -1498,8 +1498,9 @@ class ZstdDecompressionWriter(object):
 
 
 class ZstdDecompressor(object):
-    def __init__(self, dict_data=None):
+    def __init__(self, dict_data=None, max_window_size=0):
         self._dict_data = dict_data
+        self._max_window_size = max_window_size
 
         dctx = lib.ZSTD_createDCtx()
         if dctx == ffi.NULL:
@@ -1789,6 +1790,13 @@ class ZstdDecompressor(object):
 
     def _ensure_dctx(self, load_dict=True):
         lib.ZSTD_DCtx_reset(self._dctx)
+
+        if self._max_window_size:
+            zresult = lib.ZSTD_DCtx_setMaxWindowSize(self._dctx,
+                                                     self._max_window_size)
+            if lib.ZSTD_isError(zresult):
+                raise ZstdError('unable to set max window size: %s' %
+                                ffi.string(lib.ZSTD_getErrorName(zresult)))
 
         if self._dict_data and load_dict:
             zresult = lib.ZSTD_DCtx_refDDict(self._dctx, self._dict_data._ddict)